refactor: import components from internal package
