+++ src/components/__init__.py
@@ -1,0 +1,7 @@
+# __init__.py
+from components.suggest_box import SuggestBox
+from components.tab import EditorTab
+from components.windows import Window
+from components.display_manager import DisplayManager
+
+__all__ = ["Window", "SuggestBox", "EditorTab", "DisplayManager"]
